Introduce pipelining functionality
Supplying the output of one command as the input to another is an
idiomatic way of chaining commands together. This "pipeline" pattern is
applicable in a variety of use cases.
This change introduces pipeline execution functionality to be able to
chain arbitrary commands by connecting their stdin and stdout file
descriptors.