Wait for all processes of a pipe and clean them up
Previously, after pipeline creation we waited for the last process in
the pipe to terminate. This is sufficient in the sense that we will not
return too early. However, whenever a process terminates it stays around
in a zombie mode in which some of the reserved resources are not yet
freed. Among other things, the process' error code is contained in the
data structures. In order to clean up this zombie and, with that,
release the remaining allocated resources, a wait system call has to be
performed.
This change adjusts the pipeline function to return a list of process
IDs instead of a single one from the last process started. A subsequent
wait is performed on each of them in order to allow for proper cleanup.
In addition, this change will cause an exception to be raised if *any*
of the processes in the pipe terminated with a failure. This behavior is
different from that exhibited by bash (and possibly other shells) where
only the last process in the pipeline can cause the pipeline to fail. It
is still to be determined if this behavior is safe and preferred.