Introduce generic redirection facilities
Upcoming changes make it a requirement to be able to feed input to a
program which means we must be able to write data to stdin and redirect
the new process' stdin appropriately.
The current process of reading from stdout really only works when we
have data coming in from a single channel because we block on the read.
Due to this fact, it would not be possible to service stdin as well
because we might deadlock.
This change refactors the redirection and data retrieval approach. For
data retrieval we now rely on the poll functionality. This way, we get
informed if new data is available and only react based on such an event.
Furthermore, this change removes the 'executeAndRead' and
'pipelineAndRead' functions since they were too limited in what could be
achieved. Instead, we only have 'execute' and 'pipeline' but we pass in
parameters that indicate whether we have data to supply to stdin or
whether we want to read from stdout or stderr.
The defer functionality from the cleanup module is harnessed to always
free resources and invoke cleanup functions, even in the face of
exceptions.