Include stderr output in exception in case spawned process fails
Until now we could see if a program that was run using the execute or
pipeline functionality failed but we did not see the error message it
printed out to standard error (if any). For production use this behavior
is unacceptable as we need all the information we can possibly acquire
once something went wrong.
With this change the output of a failed program's standard error is
embedded into the raised ChildProcessException exception. For this
embedding to be effective, the user must have specified to read from
stderr in the first place -- we fall back to the previous behavior of
only reporting the program that was run in conjunction with the error
code otherwise.
Moreover, this change changes the default parameter read_err of the
execute and pipeline functions to true and, thus, enables the reading of
the error output by default. In the execute tests, however, we
explicitly disable this parameter to preserve the behavior we had until
now to be able to better test the different redirections in isolation
without polluting the code too much.