Iron out waiting for pipe processes and error reporting
A previous change introduced a failure model in which *any* error in a
pipeline of commands is treated as a significant failure and is reported
to the user in the form of an exception. However, it was unclear whether
this approach is safe under all circumstances since bash, for instance,
behaves differently in the sense that only the last process in a
pipeline can cause the pipeline to fail, i.e., intermediate failures are
ignored.
With this change the previously introduced model is kept and considered
the safer option. The main difference between the fail-any and fail-last
approaches is that the latter assumes that processes include error
correcting information in the stream being processed, e.g., at the very
least an end-of-stream marker that would detect short input. I argue
that this is not necessarily the case. Consider a pipe that includes a
simple transformation program such as 'tr' or even a more complex one
that encrypts (on a block basis) the incoming stream. In either case, no
knowledge about the expected amount of data can be assumed and neither
can a special input format be employed, and the entire pipeline would
succeed even if one of the previous input producing or input processing
commands failed.
Furthermore, this change addresses a potential zombie process leak in
the face of a failure in all but the last process of a pipeline. Until
now, an exception would be raised which stopped the wait for further
child processes, leaving unwaited processes dangling. This change
adjusts the algorithm to first wait for all children and only afterwards
raise an exception in case one of the programs failed.