Fix intermittent excessive read test case failures
In the past we saw intermittent test failures for the excessive read
test case where not all data written into a pipeline was being read from
it afterwards.
Investigation has shown that the reason for the data loss was that we
received an event combination of POLLIN and POLLHUP. In such a case we
read one buffer size of data (because of POLLIN) and closed the pipe
afterwards (because of POLLHUP). However, we might actually have
received more data than can fit into the buffer but we will never get
informed about that fact because we already removed the file descriptor
from the poll mechanism and will, thus, never receive follow up POLLIN
events.
This change fixes the issue by reading all available data in the case a
POLLHUP was sent in conjunction with a POLLIN.