Report pipe related errors to the user
Until now we silently ignored errors on a pipe and closed it. However,
previously we saw that we are not immune to retrieval of incomplete
data in the face of early pipe closures or (potentially) pipe connection
errors.
This change adjusts the behavior for pipe errors reported by poll, i.e.,
POLLERR and POLLNVAL. We now raise an exception to signal this error to
the user such that he/she is made aware of potentially incomplete data
that was received (or sent). This approach is considered more safe than
just ignoring errors and closing the pipe because we do not know if the
processes have error correction built into their communication or not.