Add test to verify we can read output from child processes
Some programs fork in order to perform a certain task and have each
child process print a result or something similar to one of the default
file descriptors. If such a program is started using out execution
infrastructure we want to be able to read the output as produced by any
forked off processes.
This change adds a test case that verifies that we can correctly
retrieve this output from a forked off child process.