Replace formatPipeline function with formatCommands
A subsequent change will introduce support for so called "springs" --
an arrangement of commands that, similar to a pipeline, is executed in a
certain way.
At the level relevant for this change, a spring is represented as just
another list of commands inside a pipeline. That is, one can think of it
as a single command being replaced by a list of commands. This list, in
turn, is executed in a serial manner with the output being accumulated
into a single destination (which could potentially be a following
pipeline of commands).
This change prepares the command formatting function for the
introduction of this new execution model. To that end, 'formatPipeline'
is renamed to 'formatCommands' and is able to properly format almost
arbitrary sets of commands with various levels of list nesting (for
instance, a single command has a maximum nesting level of 0 [or 1,
depending on your point of view], a pipeline a depth of one, a spring
one or two). Note that strictly speaking this special list making up a
spring is only allowed to be the first command in the pipeline but this
restriction is not enforced for the formatting of commands, making it
possible to even print invalid sets of commands (which, in theory, could
come handy for diagnosis tasks).