Introduce "spring" support
The execution functionality we got so far allows for running a single
command as well as pipelines of commands and redirecting the input and
output as desired.
One thing that is missing for supporting a greater variety of execution
patterns is the accumulation of output from various programs into a
single output stream or file descriptor.
This change adds support for what I refer to as a spring: a sequence of
commands that do not expect any input but generate output. Note that the
term sequence implies the preservation of the order (achieved by
executing them serially), meaning the output stream is still predictable
and not arbitrarily intermingled as would be the case if the commands
were run in parallel.
There may be more ways than adding a new execution primitive to achieve
this execution pattern. Another approach is the introduction of a script
built on top of the existing execute functionality. However, such an
approach introduces additional complexity because the existing 'execute'
primitive always works synchronously, meaning the function will return
only after the process finished its work. However, in order to not stall
the process sequence, it is important to start the entire command set
involved in the output processing. To that end, the means of
asynchronicity need to be introduced.
Another approach is the introduction of a new primitive which performs
the desired task. This patch adds such functionality. In essence, in
addition to the one dimensional command (a list of strings) and the two
dimensional pipelines (a list of commands, i.e., a list of a list of
strings), a spring is a three dimensional construct: a list of commands,
similar to a pipeline, with the first "command" being a list of commands
in itself. As explained, the commands in the first element are run
sequentially, with each (sub) command intended to generate output. The
output of all those (sub) commands is accumulated and then passed to the
command represented by the second element of the command list and from
there on we are back at the execution scheme of a pipeline (meaning each
commands output acts as input to the subsequent one).