Do not use lambda expressions for function deferral
Recent changes to the cleanup module made it possible to directly pass
parameters to a deferred function without the need for creating a lambda
expression first. Using direct parameters can avoid hard to detect bugs
in some scenarios where the parameter binding rules are not obvious at
first glance (such as "for" loops) and where otherwise introduction of
an additional function would be required.
This change adjusts the execute module to make use of this option.