Fix ordering bug with respect to spring pids
When a command executed as part of a spring fails, the produced error
message might refer to a different command (that actually succeeded
execution). The cause for this issue is the insertion of the PID of a
spawned process in the wrong location in the final list of PIDs of
processes started.
This change addresses the issue. A test will be added in a subsequent
change.