Fix bug in status reporting for springs
It is possible that a spring command fails but the reported status
falsely reports success (i.e., is zero). The problem is that we detected
failure and remembered the command that failed. Then we went on waiting
for all other commands to finish as well. However, the status code of
the failed process was lost and we picked another one from a potentially
successful process.
This change fixes the issue by passing through the status of a failed
process along with its name and arguments. A test will be added as part
of a subsequent change.