Rework status code handling
Reporting of status codes was not correct. We directly used the status
as returned by 'waitpid' which might be shifted version of the actual
return code. This issue was the result of wrong documentation of the
'waitpid' call on Python's side.
This patch uses the WIF* functions to check for the status condition of
a waited for process and the corresponding functions for the actual
extraction of the status code.