Add signal handling support
POSIX defines various signals that can be send to a process, some of
them exiting it. By convention, the exit code of the process in such a
case should be the negated signal number.
This change adds support for handling exiting signals correctly. A test
verifies that we conform to the behavior exhibited by the 'subprocess'
module.