Add support for passing in custom environments
It is sometimes necessary to supply a new set of environment variables
to a process without modifying the current process'. Or, if isolation is
a concern, one might want to start a new process with an empty
environment.
This change adds support for custom environments, enabling those two use
cases (and more). The major execution functions (execute, pipeline, and
spring) are extended with a new 'env' parameter that can contain an
environment. If None is used (the default) the current process'
environment will be inherited. If, on the other hand, a dict is supplied
it will be used as the environment for the new process instead
(that includes the empty dict which effectively clears the environment).