Do not inherit ProcessError from OSError (or a derived class)
Upon introduction of the ProcessError class we decided to derive it from
OSError (more precisely, ChildProcessError) in order to fit into the
same class hierarchy. However, because we introduced and used custom
attributes we are arguably misusing OSError here. The problem becomes
apparent when clients want to use one of OSError's attributes (such as
'filename' or 'strerror') which were simply not set. No problems occur
if only the textual representation (i.e., the str functionality) is
used.
With this change we decided to avoid such problems which are a potential
source of confusion. Instead of inheriting from ChildProcessError we
decided to use RuntimeError as the base class, which does not have such
attributes.