Properly strip stderr input in ProcessError
In case of a process returning an error code (that is, having an exit
status not equal to zero) we raise a ProcessError. In case we read the
stderr output from the process we supply this to the exception to raise
which in turn conveys it to the user. The problem here is that the
stderr output does not necessarily conform to a uniform format. In
particular, it can happen that superfluous whitespaces are included in
the input (think, trailing newlines) which distort the final output as
they will be included directly.
With this change we have the ProcessError class strip the stderr input
from any leading or trailing whitespaces.