Add note about child process behavior in case of death of the parent
When a process is forked off it becomes the child of its parent.
However, this relationship does not imply a strict ownership. In
particular, if the parent process dies the child will become an orphan
and will not be killed. This behavior can be influenced via the
PR_SET_PDEATHSIG prctl (see prctl(2)).
This change adds a note to the module to make this behavior clear to
clients of the module. There were three reasons for not allowing users
to influence this behavior from within the execute module:
  1) This behavior is rarely relevant in scenarios where the execute
     module is envisioned to be used.
  2) As of Python 3.4 there are no bindings for the prctl system call
     which means either a dependency to a non-standard module would have
     to be pulled in or custom bindings be created (which is a
     non-trivial effort).
  3) Since the PR_SET_PDEATHSIG setting is specified on a per-client
     basis, clients (i.e., started processes) are (already) free to set
     this flag. It is possible (although not necessarily always the
     case) that processes themselves know more about their potential
     parent and whether it makes sense to be informed (or torn down) in
     case the parent dies.