Only return output data if asked for
When originally designing the command execution APIs the stance was that
we have those functions unconditionally return the data we read from
both output streams, stdout and stderr, so as to allow for uniform
handling by clients (they can always read a 2-tuple of the stdout and
stderr data, the strings just might be empty). Along with uniformity,
another factor influencing this decision was the expectation that stderr
is set most of the time anyways in order to be able to provide a proper
error description in case of execution failure.
While the latter condition still holds, it was also found that those few
clients that conditionalize the stderr retrieval have some unattractive
boilerplate code in them.
To that end, this patch changes our approach to returning data by only
returning it for the streams asked for. That is, if no stdout data was
requested, no stdout data will be returned. The same holds for stderr.
If no data is to be read at all the (implicit) return value is None. If
data from both streams is to be read, nothing changes for clients.
Because Python has strong checks on the unpacking of tuples, the belief
is that by using this approach we force clients into being more pedantic
about from which streams to read and whether or not to evaluate the
data.