Do not return anything when redirecting to files
When we redirect an output stream into a file we, naturally, cannot
provide the user with the data produced on this stream and redirected
into the file unless we want to put in place some expensive
interception mechanism.
To that end, this change adjusts the return value for those cases to
be None instead of the supplied input byte sequence (typically b"").