Properly set the ProcessError's stderr member
Currently, when the stderr parameter to any of the execution functions
is b"" and the started process fails, the ProcessError's stderr member
is set to None is there was no actual output made by the process.
Arguably, this fact violates the principle of least astonishment because
this behavior is not in line with the respective return values of the
execution functions (where the return value is b"" if stderr was b"" and
no additional output was made to the standard error stream).
This change fixes the problem by setting the stderr attribute with the
same logic the return value is determined.