Fix error reporting on exec* failures
As it turns out all variations of the execute functionality suffer from
a problem when an exception is raised after a fork but before the exec
replaced the process image. Such an exception does not necessarily
result in a non-zero exit status of the program (noticed in particular
in an interactive interpreter) and even if an error is indicated
correctly, a very generic ProcessError is raised as a result in the
parent.
This change fixes this problem by explicitly catching any such
exceptions and encoding them for proper consumption by the parent. The
parent in turn is adjusted to attempt to parse the exception from stderr
and raising an exception of the same type.