Embed file name in case of FileNotFoundError exec failure
When execution of a non-existent program is attempted, a
FileNotFoundError will be raised. This error, however, only conveys the
fact that a file was not found, not which file it was. In many cases
(especially in larger code bases) this information is crucial for
debugging a problem more quickly.
This change adjusts the logic to actually embed the file name into the
exception. As it turns out the raised FileNotFoundError -- a subclass of
OSError -- already accepts an optional 'filename' parameter. It is just
unset when the error originates in one of the exec* functions. So with
this change we correct that by correctly supplying this argument.