Fix command formatting for spring related failures
Error reporting in springs has been a special case since its inception,
because of the execution infrastructure we have for them. Unfortunately,
this special casing also led to a subtle bug off by one bug: if the last
command in the "spring" part of a spring fails (that is, the b in the
spring represented as [[a, b], c]), we will format said spring part as
the failing command, leading to unintuitive output such as:
  > [Status 1] (/bin/true + /bin/false)

With this change we fix this problem by flattening the spring before
indexing into it to retrieve the individual commands.