Use dedicated pipe for exception transmission
With a recent change we piggy backed on using stderr for relaying
exception information from the child to the parent after a fork.
Overloading the usage of this communication channel is not clean design
as we mix different concerns and in fact cannot work if clients decided
to redirect a child process' stderr to their own, in which case this
exception information will show up.
To fix this issue this change introduces a pipe that is used as the
communication channel for this type of information from the child to the
parent.