Fix decoding error in case of multiple process failures
Unfortunately, with the move towards communicating exec* failures from
sub processes to the parent we introduced a problem when multiple
children encounter an error: all of them will emit their exception
information to the parent but decoding at the parent will fail because
of excess data (more than one encoded exception).
No proper solution to this problem has been found as of now, but the
workaround is to emit exceptions separated by a new line symbol. On the
decoding side we now only look at the very first line. Unfortunately,
this method, however is not race free: it is possible that two child1
which is spawned first emits its failure data later than child2, which
is spawned later. We have to punt on this problem for now.